requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/kidig/api.git@a0c97026fbd00cb7be9dcb796135779b4810cef5#egg=api
-e git+https://github.com/kidig/django-robust.git@586589ab159002461f98422b9831b2f89ee0a10c#egg=django_robust
colorlog==2.10.0
diff-match-patch==20121119  # via django-import-export
django-happymailer==0.1.8
django-import-export==0.4.5  # via django-happymailer
django-object-actions==0.8.2
django==1.10.3
easy-thumbnails==2.3
factory-boy==2.7.0
fake-factory==0.7.2       # via django-happymailer, factory-boy
html2text==2016.5.29      # via django-happymailer
jsonschema==2.5.1
pillow==3.4.2             # via easy-thumbnails
psycopg2==2.6.2
py==1.4.31                # via pytest
pytest-django==3.1.2
pytest==3.0.4
python-dateutil==2.6.0    # via fake-factory
pyyaml==3.12
requests==2.12.1
schedule==0.3.2
six==1.10.0               # via django-happymailer, fake-factory, python-dateutil
tablib==0.11.2            # via django-import-export
trafaret==0.7.5
uwsgi==2.0.14
